wim-mngr/requirements.txt
# Mongo databse
PyMongo

# Kafka message bus
kafka-python